MAJOR Add support to UserScript and vote count
This is a final release that could potentially be updated if needed. In
fact, it might need as this only supports a single pool at a the momment
per userscript added to TamperMonkey or others.

It can count the number of votes and shows this counting on the screen.
This counting is done by cookies.

Part of the code was taken from gfvote. See gfvote for more information.
Link: https://github.com/jfcandidofilho/gfvote